Split standard implementation for dynamic assets
