+++ bevy_asset_loader/src/loading_state/dynamic_asset_systems.rs
@@ -1,7 +1,6 @@
-use crate::dynamic_asset::{
-    DynamicAssetCollection, DynamicAssetCollections, DynamicAssets, StandardDynamicAssetCollection,
-};
+use crate::dynamic_asset::{DynamicAssetCollection, DynamicAssetCollections, DynamicAssets};
 use crate::loading_state::{InternalLoadingState, LoadingAssetHandles};
+use crate::standard_dynamic_asset::StandardDynamicAssetCollection;
 use bevy::asset::{AssetServer, Assets, LoadState};
 use bevy::ecs::change_detection::ResMut;
 use bevy::ecs::schedule::State;
